[#60] feat: 예외 처리 메서드 클래스로 분리
